pyproject.toml
[project]
name = "datamining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
